# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: ParDespue API
  description: API para gestionar usuarios, roles, cápsulas, contenidos, notificaciones y más.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Servidor local
paths:
...
  /contents:
    get:
      summary: Obtener todos los contenidos
      tags: [Contents]
      responses:
        '200':
          description: Lista de contenidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Content'
        '500':
          description: Error interno del servidor
    post:
      summary: Crear contenido (archivo) asociado a una cápsula
      tags: [Contents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentInput'
      responses:
        '201':
          description: Contenido creado exitosamente
        '400':
          description: Error de validación
        '500':
          description: Error interno del servidor
  /contents/capsule/{capsuleId}:
    get:
      summary: Obtener todos los contenidos de una cápsula
      tags: [Contents]
      parameters:
        - name: capsuleId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Lista de contenidos de la cápsula
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Content'
        '500':
          description: Error interno del servidor
  /contents/{id}:
    delete:
      summary: Eliminar un contenido por ID
      tags: [Contents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Contenido eliminado
        '404':
          description: Contenido no encontrado
        '500':
          description: Error interno del servidor
  /contents/by-capsule/{capsuleId}:
    delete:
      summary: Eliminar todos los contenidos de una cápsula
      tags: [Contents]
      parameters:
        - name: capsuleId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Contenidos eliminados
        '500':
          description: Error interno del servidor
...
  /subscriptions/my-plan:
    get:
      summary: Obtener el plan actual del usuario autenticado
      tags: [Subscriptions]
      responses:
        '200':
          description: Plan actual del usuario
        '401':
          description: No autenticado
        '500':
          description: Error interno del servidor
  /subscriptions/stripe-prices:
    get:
      summary: Obtener precios de Stripe para los planes
      tags: [Subscriptions]
      responses:
        '200':
          description: Lista de precios de Stripe
        '500':
          description: Error interno del servidor
  /subscriptions/create-checkout-session:
    post:
      summary: Crear sesión de pago de Stripe para suscripción Premium
      tags: [Subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan: { type: string }
                billing: { type: string }
                priceId: { type: string }
      responses:
        '200':
          description: URL de Stripe para el pago
        '401':
          description: No autenticado
        '500':
          description: Error interno del servidor
  /subscriptions/activate-premium:
    post:
      summary: Activar plan premium después del pago
      tags: [Subscriptions]
      responses:
        '200':
          description: Premium activado correctamente
        '401':
          description: No autenticado
        '500':
          description: Error interno del servidor
  /subscriptions/renew/{id}:
    post:
      summary: Renovar una suscripción
      tags: [Subscriptions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                months: { type: integer }
      responses:
        '200':
          description: Suscripción renovada
        '401':
          description: No autenticado
        '500':
          description: Error interno del servidor
  /subscriptions/cancel/{id}:
    post:
      summary: Cancelar una suscripción
      tags: [Subscriptions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Suscripción cancelada
        '401':
          description: No autenticado
        '500':
          description: Error interno del servidor
...
  /auth/register:
    post:
      summary: Registrar un nuevo usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                profile_picture:
                  type: string
                  format: binary
      responses:
        '201':
          description: Usuario registrado correctamente
        '400':
          description: Error de validación
        '500':
          description: Error interno del servidor
  /auth/login:
    post:
      summary: Iniciar sesión
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
                remember: { type: boolean }
      responses:
        '200':
          description: Login exitoso
        '400':
          description: Usuario o contraseña incorrectos
        '403':
          description: Usuario no verificado o baneado
        '500':
          description: Error interno del servidor
  /auth/verify/{token}:
    get:
      summary: Verificar cuenta de usuario
      tags: [Auth]
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Usuario verificado correctamente
        '404':
          description: Token inválido o expirado
        '500':
          description: Error interno del servidor
  /auth/resend-verification:
    post:
      summary: Reenviar correo de verificación
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200':
          description: Correo de verificación reenviado
        '404':
          description: Usuario no encontrado
        '400':
          description: Usuario ya verificado
        '500':
          description: Error interno del servidor
...
  /auth/reset-password:
    post:
      summary: Restablecer contraseña con token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Contraseña restablecida correctamente
        '400':
          description: Token inválido o expirado
        '500':
          description: Error interno del servidor
...
components:
  schemas:
...
    Content:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        filePath: { type: string }
        Name: { type: string }
        Capsule_ID: { type: integer }
    ContentInput:
      type: object
      properties:
        type: { type: string }
        filePath: { type: string }
        Name: { type: string }
        Capsule_ID: { type: integer }
...